requirements.txt
chromadb>=0.5.0
pyyaml>=6.0
requests>=2.31
streamlit>=1.30
